Introduce item container
