+++ src/internal/item-container/styles.scss
@@ -1,0 +1,6 @@
+.root {
+  touch-action: none;
+  position: relative;
+  height: 100%;
+  display: grid;
+}
