+++ src/item/styles.scss
@@ -39,9 +39,6 @@
   position: relative;
   height: 100%;
   display: grid;
-  &.wrapperDragging {
-    z-index: 1;
-  }
 }
 
 .body {
